Assume door is closed when it gets locked
- Moved helper methods to its own module
