+++ src/lib.rs
@@ -26,6 +26,7 @@ pub use verb::Verb;
 
 mod verb;
 mod actions;
+mod helpers;
 
 #[derive(Clone, Debug, PartialEq)]
 pub enum Adjective {
@@ -219,51 +220,6 @@ impl Object {
                 }
             }
         }
-    }
-
-    /// Returns `true` if self has another object.
-    pub fn has(&self, obj: Object) -> bool {self.matches(&has(obj))}
-    /// Returns `true` if self has not another object.
-    pub fn has_not(&self, obj: Object) -> bool {self.matches(&has_not(obj))}
-    /// Returns `true` if self is on another object.
-    pub fn is_on(&self, obj: Object) -> bool {
-        self.matches(&on(obj).into())
-    }
-    /// Returns `true` if self leans toward another object.
-    pub fn is_leaning_toward(&self, obj: Object) -> bool {
-        self.matches(&lean_toward(obj).into())
-    }
-    /// Returns `true` if self is in another object.
-    pub fn is_in(&self, obj: Object) -> bool {
-        self.matches(&in_(obj).into())
-    }
-    /// Returns `true` if self is out of another object.
-    pub fn is_out_of(&self, obj: Object) -> bool {
-        self.matches(&out_of(obj).into())
-    }
-    /// Returns `true` if self was killed by another object.
-    pub fn was_killed_by(&self, obj: Object) -> bool {self.matches(&killed_by(obj))}
-    /// Returns `true` if self killed another object.
-    pub fn killed(&self, obj: Object) -> bool {self.matches(&killed(obj))}
-    /// Returns `true` if self talked to another object.
-    pub fn talked_to(&self, obj: Object) -> bool {
-        self.matches(&Object::DidTo(Verb::Talk, Box::new(obj)))
-    }
-    /// Returns `true` if self was talked to by another object.
-    pub fn was_talked_to_by(&self, obj: Object) -> bool {
-        self.matches(&Object::WasBy(Verb::Talk, Box::new(obj)))
-    }
-    /// Returns `true` if self was moved by another object.
-    pub fn was_moved_by(&self, obj: Object) -> bool {
-        self.matches(&Object::WasBy(Verb::Move, Box::new(obj)))
-    }
-    /// Returns `true` if self moved another object.
-    pub fn moved(&self, obj: Object) -> bool {
-        self.matches(&Object::DidTo(Verb::Move, Box::new(obj)))
-    }
-    /// Returns `true` if self is opponent of another object.
-    pub fn is_opponent_of(&self, obj: Object) -> bool {
-        self.matches(&opponent_of(obj).into())
     }
 }
 
@@ -547,6 +503,19 @@ mod tests {
         assert!(room.action(&He.gives_item(She, key_to(of_type("door")))).is_err());
         room.action(&He.picks_up(key_to(of_type("door")))).unwrap();
         room.action(&He.gives_item(She, key_to(of_type("door")))).unwrap();
+
+        let i = 0;
+        let door = 1;
+        let mut room = Room::new(vec![I, of_type("door")]);
+        assert!(!room.objects[door].matches(&Locked.into()));
+        assert!(!room.objects[door].matches(&Closed.into()));
+        room.action(&I.locks(of_type("door"))).unwrap();
+        assert!(room.objects[door].matches(&Locked.into()));
+        // Assume door is closed when it gets locked.
+        assert!(room.objects[door].matches(&Closed.into()));
+        assert!(room.objects[i].locked(of_type("door")));
+        // Only because I locked the door, does not mean I closed it.
+        assert!(!room.objects[i].closed(of_type("door")));
     }
 
     #[test]
